♻️ (webapp-svelte) Change structure
