+++ webapp-svelte/package.json
@@ -9,6 +9,7 @@
     "awesome-notifications": "^3.1.1",
     "bootstrap": "4",
     "dompurify": "^2.2.7",
+    "eventemitter3": "^4.0.7",
     "lodash": "^4.17.21",
     "marked": "^2.0.1",
     "page": "^1.11.6",
