+++ webapp-svelte/src/components/Game/GameSidebar/index.ts
@@ -1,0 +1,5 @@
+export { default as GameLog } from "./GameLog.svelte";
+export { default as GameNotes } from "./GameNotes.svelte";
+export { default as GamePreferences } from "./GamePreferences.svelte";
+export { default as GameSettings } from "./GameSettings.svelte";
+export { default as ReplayControls } from "./ReplayControls.svelte";
