+++ webapp-svelte/src/components/Game/index.ts
@@ -1,0 +1,4 @@
+export * from "./GameSidebar";
+export { default as GameSidebar } from "./GameSidebar.svelte";
+export { default as OpenGame } from "./OpenGame.svelte";
+export { default as StartedGame } from "./StartedGame.svelte";
