+++ webapp-svelte/src/components/Layout/Footer.svelte
