+++ webapp-svelte/src/components/Layout/GameListSidebar.svelte
