+++ webapp-svelte/src/components/Layout/Sidebar.svelte
