+++ webapp-svelte/src/components/Layout/index.ts
@@ -1,0 +1,4 @@
+export { default as Appbar } from "./Appbar.svelte";
+export { default as Footer } from "./Footer.svelte";
+export { default as GameListSidebar } from "./GameListSidebar.svelte";
+export { default as Sidebar } from "./Sidebar.svelte";
