+++ webapp-svelte/src/components/User/index.ts
@@ -1,0 +1,5 @@
+export { default as PreferencesChooser } from "./PreferencesChooser.svelte";
+export { default as UserElo } from "./UserElo.svelte";
+export { default as UserGames } from "./UserGames.svelte";
+export { default as UserGameSettings } from "./UserGameSettings.svelte";
+export { default as UserPublicInfo } from "./UserPublicInfo.svelte";
