+++ webapp-svelte/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Game";
+export { default as GameList } from "./GameList.svelte";
+export * from "./Layout";
+export * from "./User";
